additional change to prevent non error status message

+++ getpackages.sh
@@ -3,7 +3,11 @@
 # dpkg-query -l | grep -v "^rc" > packages
 # echo checking installed packages
 
-rm packages
+if [ -e packages ]
+then
+	rm packages
+fi
+
 touch packages
 
 GET http://$1/config/$2/packages.txt > packages
